feat: id별tradeList fetch 기능 추가, trade-store pinia추가 #8
